feat:create handler and service recover_transaction, create default module to inject the postgre connection, create operations_types.go to verify if the operations_type in transaction is valid

+++ internal/app/controle_de_transacao/provider/postgre/model/transaction.go
@@ -10,7 +10,7 @@ type Transaction struct {
 	ID              int       `pg:"id"`
 	AccountID       int       `pg:"account_id"`
 	OperationTypeID int       `pg:"operation_type_id"`
-	Amount          float32   `pg:"amount"`
+	Amount          float64   `pg:"amount"`
 	EventDate       time.Time `pg:"event_date"`
 }
