+++ internal/app/controle_de_transacao/server/rest/model/response/transaction.go
@@ -9,7 +9,7 @@ type Transaction struct {
 	ID              int       `json:"id"`
 	AccountID       int       `json:"account_id"`
 	OperationTypeID int       `json:"operation_type_id"`
-	Amount          float32   `json:"amount"`
+	Amount          float64   `json:"amount"`
 	EventDate       time.Time `json:"event_date"`
 }
